Remove unnecessary usage of `{{ site.url }}`
It turns out that doing this is not necessary, and it is way easier and more convenient to omit it and use plain `{% post_url %}`.

I _think_ the reason I started doing this was because, awhile back, I tried out having a mailing list for my blog posts. I was using MailChimp's rss-to-email plugin, and they did not correctly resolve relative URLs for the emails. Anyway, I'm not doing that now so these are no longer needed. (Also, images have always used relative links.)

If I ever want to return to this, I can use my new Liquid Tag, `post_absolute_url`, which will be an easy find-and-replace.

See: #141, #142