Actually fix bug in recently updated post list
This was thought to be fixed by 1b4a0ea0, but actually was not fixed.

The issue was that `updated_posts` was getting truncated and showing only the **first** 30 (`site.feeds.limit`) updated posts, not the _most recent_ 30 updated posts.

The change in 1b4a0ea0 fixed the limiting issue, but introduced a new bug where the posts were not sorted by `date-updated`.

The underlying problem is that using `reversed` with `limit:` does not work as expected. I'm not sure if this is a bug, or expected behavior.

The `updated_posts` variable is filtered and sorted by `post.date-updated`. Previously, this loop used `limit: site.feeds.limit` to limit the total number of posts shown, similar to how `post_list` is handled below.

However, even though we are using `reversed` to sort by the most recently updated post first, `limit:` appears to limit the count in-order using the *non-reversed* array. Thus, you end up with the first `N` updated posts rather than the `N` most recent.

To fix, we need to manually check the loop index against our feed limit, then break out of the loop.